Change the type of variables

+++ kurdikova.elizaveta/I2/GetMaxCountSameNumbersSeq.cpp
@@ -2,8 +2,8 @@
 
 int getMaxCountSameNumbersSeq(const int* arr, size_t size)
 {
-  int maxCount = 1;
-  int currentCount = 1;
+  size_t maxCount = 1;
+  size_t currentCount = 1;
   for (size_t i = 0; i < size - 1; ++i) {
     if (arr[i] == arr[i + 1]) {
       currentCount++;
